refactor(backend): modularize and refactor code
