--- credentials.json
@@ -1,1 +1,0 @@
-{"installed":{"client_id":"30535908325-m4a18voskun7gd2lg646q1gk3e26f8t0.apps.googleusercontent.com","project_id":"eventup-1598682034730","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"AF3Iy5n8RPeWWlP95E4B1kSk","redirect_uris":["urn:ietf:wg:oauth:2.0:oob","http://localhost"]}}
