+++ frontend/package.json
@@ -3,22 +3,22 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.8",
-    "axios": "^0.20.0",
+    "@testing-library/jest-dom": "^5.11.8",
+    "@testing-library/react": "^11.2.3",
+    "@testing-library/user-event": "^12.6.0",
+    "axios": "^0.21.1",
     "bootstrap": "^4.5.3",
     "js-cookie": "^2.2.1",
     "jsonwebtoken": "^8.5.1",
-    "qrcode.react": "^1.0.0",
-    "react": "^16.14.0",
-    "react-bootstrap": "^1.3.0",
-    "react-dom": "^16.14.0",
-    "react-popper": "^2.2.3",
+    "qrcode.react": "^1.0.1",
+    "react": "^17.0.1",
+    "react-bootstrap": "^1.4.3",
+    "react-dom": "^17.0.1",
+    "react-popper": "^2.2.4",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "^3.4.3",
-    "react-toastify": "^6.0.9",
-    "reactstrap": "^8.6.0"
+    "react-scripts": "^4.0.1",
+    "react-toastify": "^6.2.0",
+    "reactstrap": "^8.8.1"
   },
   "scripts": {
     "start": "react-scripts start",
